refactor: move RaftServer interface and GoRaftServer to package raft
