+++ weed/command/server.go
@@ -12,6 +12,7 @@ import (
 
 	"weed/glog"
 	"weed/pb/master_pb"
+	"weed/raft"
 	weed_server "weed/server"
 	"weed/storage"
 	"weed/util"
@@ -163,7 +164,7 @@ func runServer(cmd *Command, args []string) bool {
 				mPeers[string(peer)] = peer
 			}
 
-			raftServerOption := &weed_server.RaftServerOption{
+			raftServerOption := &raft.RaftServerOption{
 				Peers:             mPeers,
 				ServerAddr:        myMasterAddress,
 				DataDir:           *metaFolder,
